# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Nexvault — Digital Banking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /auth/login:
    post:
      operationId: login
      summary: Login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify email OTP
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailResponse"
...
  /auth/otp/request:
    post:
      operationId: requestOtp
      summary: Request step-up OTP
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Onboarding
...
  /accounts:
    get:
      operationId: getAccounts
      summary: List accounts
      tags: [accounts]
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      summary: Create account
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /accounts/{id}:
    get:
      operationId: getAccount
      summary: Get account
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}/transactions:
    get:
      operationId: getAccountTransactions
      summary: Account transactions
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [credit, debit]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"

  # Cards
  /cards:
    get:
      operationId: getCards
      summary: List cards
      tags: [cards]
      responses:
        "200":
          description: Cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      summary: Issue card
      tags: [cards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Card issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      summary: Get card
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/freeze:
    put:
      operationId: freezeCard
      summary: Freeze card
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card frozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/unfreeze:
    put:
      operationId: unfreezeCard
      summary: Unfreeze card
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card unfrozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"

  # Payments
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, accountType]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        accountType:
          type: string
          enum: [individual, business]
        firstName:
          type: string
        lastName:
          type: string
    RegisterResponse:
      type: object
      required: [userId, nextStep]
      properties:
        userId:
          type: string
        nextStep:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    VerifyOtpRequest:
      type: object
      required: [userId, code]
      properties:
        userId:
          type: string
        code:
          type: string
    VerifyEmailResponse:
      type: object
      required: [verified, nextStep]
      properties:
        verified:
          type: boolean
        nextStep:
          type: string
...
    RequestOtpRequest:
      type: object
      required: [purpose]
      properties:
        purpose:
          type: string
          enum: [step_up, phone_verify, password_reset]
...
    Account:
      type: object
      required: [id, currency, balance, availableBalance, status, createdAt]
      properties:
        id:
          type: string
        currency:
          type: string
          enum: [USD, EUR, GBP, SGD, AED]
        iban:
          type: string
        accountNumber:
          type: string
        sortCode:
          type: string
        balance:
          type: number
        availableBalance:
          type: number
        status:
          type: string
          enum: [active, frozen, closed]
        createdAt:
          type: string
          format: date-time
...
    CreateAccountRequest:
      type: object
      required: [currency]
      properties:
        currency:
          type: string
          enum: [USD, EUR, GBP, SGD, AED]
    Transaction:
      type: object
      required: [id, type, amount, currency, description, status, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum: [credit, debit]
        category:
          type: string
          enum: [transfer, card, fee, fx, interest]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        reference:
          type: string
        counterparty:
          type: object
          properties:
            name:
              type: string
            accountNumber:
              type: string
            bankName:
              type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        createdAt:
          type: string
          format: date-time
    TransactionList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
    Card:
      type: object
      required: [id, cardType, lastFour, status, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        cardType:
          type: string
          enum: [virtual, physical]
        lastFour:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        status:
          type: string
          enum: [active, frozen, cancelled, expired]
        spendLimitDaily:
          type: number
        spendLimitMonthly:
          type: number
        createdAt:
          type: string
          format: date-time
    CreateCardRequest:
      type: object
      required: [accountId, cardType]
      properties:
        accountId:
          type: string
        cardType:
          type: string
          enum: [virtual, physical]
...
    UserProfile:
      type: object
      required: [id, email, accountType, kycStatus, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        accountType:
          type: string
          enum: [individual, business]
        kycStatus:
          type: string
          enum: [pending, in_progress, approved, rejected]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...